Add real photo upload, change obsolete modules
